requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.1.1
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9